Move setup_logging to sbot.logging
